:recycle: #1 remove implicit conversions

+++ src/core/window/update_view.cpp
@@ -13,6 +13,6 @@ void Window::updateView(const sf::Event &event)
 {
     sf::View v = _display.getView();
 
-    v.setSize(sf::Vector2f(event.size.width, event.size.height));
+    v.setSize(sf::Vector2f((float)event.size.width, (float)event.size.height));
     _display.setView(v);
 }
